tx id to u32

+++ src/error/client_error.rs
@@ -6,8 +6,8 @@ pub type ClientResult<T> = std::result::Result<T, ClientError>;
 #[derive(Clone, Debug, Error)]
 pub enum ClientError {
     #[error("Could not find transaction by id of {0}")]
-    TransactionNotFound(u16),
+    TransactionNotFound(u32),
 
     #[error("Transaction {0} was not disputed")]
-    NotDisputed(u16)
+    NotDisputed(u32)
 }
